requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
distro==1.9.0
exceptiongroup==1.2.2
groq==0.11.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
numpy==2.1.1
openai==1.45.1
pandas==2.2.2
pydantic==2.9.1
pydantic_core==2.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
scipy==1.14.1
sentence-splitter==1.4
six==1.16.0
sniffio==1.3.1
tdqm==0.0.1
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
